requirements.txt
# Core
pinecone
openai>=1.0.0
python-dotenv>=1.0.0
fastapi
python-multipart
uvicorn[standard]

# Chunking
langchain-text-splitters>=0.3.0

# File processing
pypdf>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0
markdown>=3.0.0
beautifulsoup4>=4.0.0
pdf2image
pytesseract
pillow